requirements.txt
rply>=0.7.3
pyrsistent>=0.10.1
greenlet>=0.4.5
eventlet>=0.17.2
msgpack-python>=0.4.6
typeannotations>=0.1.0
astunparse>=1.2.2
